Add condition if same operator used twice
Adds a condition that if the same operator is used twice in a row, in
order to prepare to call the equals function that will be implemented
at a later point in time.
This also triggers if any operator is used when an operator and a
second number after the operator exists.